+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ```
+* Commit: 3d0160c - Update comments
+| Author: fjrodafo - Tue, 18 Feb 2025 00:46:21 +0000
+| 
 * Commit: 74c493c - Update .yaml
 | Author: fjrodafo - Mon, 17 Feb 2025 20:43:29 +0000
 |